requirements.txt
flask==2.3.3
requests==2.31.0
beautifulsoup4==4.12.2
html2text==2020.1.16
rich==13.5.2
python-dotenv
